+++ UnitTests/DnaViewTests.swift
@@ -12,31 +12,31 @@ class DnaViewTests: FBSnapshotTestCase {
         recordMode = false
     }
 
-    private func snapshot(_ component: DnaViews) {
+    private func snapshot(_ component: DnaDemoViews) {
         FBSnapshotVerifyView(component.viewController.view)
     }
 
     // MARK: - Tests
 
     func testMissingSnapshotTests() {
-        for element in elementWithoutTests(for: DnaViews.self, testMethodPrefix: "testDnaViews") {
+        for element in elementWithoutTests(for: DnaDemoViews.self, testMethodPrefix: "testDnaDemoViews") {
             XCTFail("Not all elements were implemented, missing: \(element.rawValue)")
         }
     }
 
-    func testDnaViewsColor() {
+    func testDnaDemoViewsColor() {
         snapshot(.color)
     }
 
-    func testDnaViewsFont() {
+    func testDnaDemoViewsFont() {
         snapshot(.font)
     }
 
-    func testDnaViewsSpacing() {
+    func testDnaDemoViewsSpacing() {
         snapshot(.spacing)
     }
 
-    func testDnaViewsAssets() {
+    func testDnaDemoViewsAssets() {
         snapshot(.assets)
     }
 }
